changed qos profile of subscriber to reliable from best_effort
the color image from spot must be reliable for 15fps, so it is required that all down stream image processing nodes, like apriltag, be reliable for consistency. before change we would get maybe 1-2 tag detections for every 10 images when an apriltag is in the scene because there was a synchronization problem between camera_info and image topics. now we get all tags in the scene for each image.